Create Swap two node (linked list)
Swapping Nodes in a Linked List

You are given a singly linked list and two positions x and y. Write a function to swap the nodes at these positions. The function should update the links in the list to perform the swap. Do not simply swap the data in the nodes.

Input Format:

The first line contains an integer n, the number of elements in the linked list.
The second line contains n space-separated integers, representing the elements of the linked list.
The third line contains two integers x and y, the positions of the nodes to be swapped (0-indexed).
Output Format:

Print the linked list after swapping the nodes at positions x and y.
[Refer sample input and output]
[All text in bold refers to the input]

Sample Input 1:

Enter the number of elements:

6

Enter the elements:

10 15 12 13 20 14

Enter the positions to swap:

2 4